unify with/without docker-machine #49, #50
Use the unified logic for all platforms when docker-machine is
in use or not.

This refactoring potentially fix the issues with Docker for
Mac/Windows.
